requirements.txt
Flask==3.0.0
SQLAlchemy>=1.4.0
pytest>=7.0.0
PyYAML>=6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # For PostgreSQL support
mysql-connector-python==8.2.0  # For MySQL support
pandas>=1.3.0  # For data manipulation and comparison
colorama==0.4.6  # For colored console output 